backend/requirements.txt
fastapi
uvicorn
scikit-learn==1.7.1
joblib
numpy
pandas
pydantic
email-validator
python-dotenv
python-multipart
